Move user delete process from controller to service
Handle a ForbiddenException instead of and Unauthoriszed error

+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,8 +1,14 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.ForbiddenException;
 import com.openclassrooms.starterjwt.exception.NotFoundException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.repository.UserRepository;
+
+import java.util.Objects;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,6 +20,14 @@ public class UserService {
     }
 
     public void delete(Long id) {
+        User user = findById(id);
+
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!Objects.equals(userDetails.getUsername(), user.getEmail())) {
+            throw new ForbiddenException(
+                    "You are not allowed to delete this user.");
+        }
         this.userRepository.deleteById(id);
     }
